added custom list functionality

+++ app.js
@@ -4,7 +4,6 @@ const mongoose = require('mongoose');
 
 const app = express();
 const port = 3000;
-const workItems = [];
 
 
 // Mongoose configuration for MongoDB connection
@@ -16,6 +15,13 @@ const itemSchema = {
 };
 
 const Item = mongoose.model('Item', itemSchema);
+
+const listSchema = {
+    name: String,
+    items: [itemSchema]
+}
+
+const CustomList = mongoose.model('customList', listSchema);
 
 //Setting up app
 app.set("view engine", "ejs");
@@ -35,7 +41,7 @@ app.listen(port, function () {
 app.get('/', function (req, res) {
 
     const items = []
-    Item.find({}, (e,items) => {
+    Item.find({}, (e, items) => {
         res.render('list', {
             listTitle: "Today",
             listItems: items,
@@ -43,14 +49,37 @@ app.get('/', function (req, res) {
     });
 });
 
-// GET /work
+// GET any custom list
 
-app.get('/work', function (req, res) {
-    res.render('list', {
-        listTitle: "Work Task",
-        listItems: workItems,
-    });
-});
+app.get("/:listName", (req, res) => {
+
+    const newListName = req.params.listName;
+
+    CustomList.findOne({
+        name: newListName
+    }, (e, list) => {
+        if (!e) {
+            if (!list) {
+                const newList = new CustomList({
+                    name: newListName,
+                    items: []
+                })
+                newList.save();
+                res.render('list', {
+                    listItems: newList.items,
+                    listTitle: newList.name,
+                })
+            } else {
+                res.render('list', {
+                    listItems: list.items,
+                    listTitle: list.name
+                })
+            }
+        }
+    })
+})
+
+
 
 // GET /about
 
@@ -62,25 +91,50 @@ app.get('/about', function (req, res) {
 
 app.post('/', function (req, res) {
 
+    const listTitle = req.body.listTitle;
     const item = req.body.newItem;
     const newItem = new Item({
         name: item
     });
-    newItem.save();
-    res.redirect('/');
+
+    if (listTitle == "Today") {
+        newItem.save();
+        res.redirect('/');
+    } else {
+        CustomList.findOne({
+            name: listTitle
+        }, (e, list) => {
+            list.items.push(newItem);
+            list.save();
+            res.redirect("/" + listTitle);
+        })
+    }
 });
 
 //POST /delete
 
 app.post('/delete', function (req, res) {
-    const deletedItemID = req.body.checkbox
-    Item.findByIdAndDelete(deletedItemID, e => {
-        if (e) {
-            console.log(e);
-        } else {
-            console.log("Deleted item")
-        }
-    })
-    res.redirect('/');
+
+    const deletedItemID = req.body.checkbox;
+    const listTitle = req.body.listTitle;
+    console.log(listTitle)
+
+    if (listTitle == "Today") {
+        Item.findByIdAndDelete(deletedItemID, e => {
+            if (e) {
+                console.log(e);
+            } else {
+                console.log("Deleted item")
+            }
+        })
+        res.redirect('/');
+    } else {
+        CustomList.findOneAndUpdate({name: listTitle},{$pull: {items: {_id: deletedItemID}}}, (e, list) => {
+            if (!e){
+                res.redirect('/' + listTitle);
+            }
+        });
+    }
+
 
 })
